Testing the Nextion
HMI speed control and UI has to be included in the AS03 V2. For the particular functionality, the testing for PD is under progress. Testing the individual Components are in track. As the part codes are to be uploaded - this is the display Nextion communicating with the controller using UART protocol sending predefined strings as data to the microcontroller to process and execute logic- through the serial bus.