Add files via upload
While Testing NEMA34 for AS V2, simply interfaced an ESP32 to the system of single mega and 2 stepper motors to control the enabling of the motors via 2 switches on any device with a browser to display a webpage and active internet connection. Provided the ESP32 and the device used to access the webpage are connected over the same local host or AP.
The ESP_MotorControl_webpage program is to be uploaded to an ESP32 with ssid and password same as the AP you are using. The Motor_ESP_control code has to be uploaded to an Arduino Mega2560. Maintain Serial communication between mega and ESP using hardware pins. Refer to the program for the pins. Or refer the pinout and documentation of both controllers.